requirements.txt
aiohttp==3.6.2
astunparse==1.6.3
async-timeout==3.0.1
asyncpg==0.22.0
attrs==19.3.0
braceexpand==0.1.6
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
discord.py
googletrans==3.0.0
gTTS==2.2.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.12.22
httpcore==0.9.1
httpx==0.13.3
humanize==3.1.0
hyperframe==5.2.0
idna==2.10
import-expression==1.1.4
jishaku==1.20.0.220
lavalink==3.1.2
multidict==4.7.6
Pillow==8.0.0
py-dotenv==0.1
pycparser==2.20
Pygments==2.8.1
pygount==1.2.4
PyNaCl==1.3.0
python-dateutil==2.8.1
python-dotenv==0.15.0
requests==2.24.0
rfc3986==1.4.0
six==1.15.0
sniffio==1.2.0
urllib3==1.25.10
wavelink==0.9.4
yarl==1.5.1
